requirements.txt
fastapi[all]
uvicorn
mysql-connector-python
